added image carousel in ImageDialog
