+++ client/package.json
@@ -27,6 +27,7 @@
     "react-clamp-lines": "^3.0.2",
     "react-dom": "^17.0.2",
     "react-geocode": "^0.2.3",
+    "react-material-ui-carousel": "^2.3.1",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
     "react-styleguidist": "^11.1.7",
